[master] Refactor and add CRUD functionality for categories and products
Standardize CRUD operations across categories, products, and stock. Introduced interface-based facades, mappers, and services for improved modularity and code reusability. Adjusted repository logic and exceptions handling to ensure compliance with business rules.

+++ src/main/java/br/com/menchitech/ms_pdv_bms_restaurant/persistence/entity/EstoquePersistenceEntity.java
@@ -14,6 +14,7 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.hibernate.annotations.UuidGenerator;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @Data
@@ -28,15 +29,15 @@ public class EstoquePersistenceEntity {
      @GeneratedValue
      @UuidGenerator
      @Column(name = "id", updatable = false, nullable = false)
-     private UUID id;
+     private UUID                            id;
 
-     private String                      nomeEstoque;
+     private String                          nomeEstoque;
 
      @OneToOne(fetch = FetchType.LAZY)
      @JoinColumn(name = "produto_id", nullable = false)
-     private ProdutoPersistenceEntity produto;
+     private ProdutoPersistenceEntity        produto;
 
-     private Integer                     quantidadeGuardada;
-     private Double                      precoSomaVenda;
-     private Double                      precoSomaCusto;
+     private Integer                         quantidadeGuardada;
+     private BigDecimal                      precoSomaVenda;
+     private BigDecimal                      precoSomaCusto;
 }
